Fix devDependencies and mention of global packages
It is not necessary to have those prerequisite packages installed
globally. When running npm scripts, the bin (e.g. browserify) is
found automatically in /node_modules/.bin/, before it tries to
find it globally. This commit also includes browserify as a
devDependency, since it was the only one missing in order for
`build_all` to run.