fix(operators): reorder signature of resultSelectors
This commit swaps around the arguments of resultSelector functions
of operators that internally create an Observable<Observable<T>>,
such as switchMap, mergeMap, and concatMap.

BREAKING CHANGES:
The function signature of resultSelectors used to be (innerValue,
outerValue, innerIndex, outerIndex) but this commits changes it to
be (outerValue, innerValue, outerIndex, innerIndex), to match
signatures in RxJS 4.