style(src): add lint tasks for TypeScript and JS
Establish a code style using tslint (for src/) and eslint (for perf/
and spec/). The existing code style was approximately inferred by
analyzing which style was the most common. The lint tasks should also
run in Travis. An established code style is important for a large
project such as RxJS which expects multiple contributions from the
community, in order to avoid pollution of feature-related commits with
style refactors.