fix(ajax): remove unnecessary onAbort handling
Since the xhr object is closed over, nothing external will be able to abort it. At best this code will
be called and then swallowed because the Subscriber is already unsubscribed and will not allow the error to
propogate through the Subscriber chain. It is basically cruft and can be removed