docs(Observable): fix typos
I'm not sure I made correct changes, but it definitely feels like something is missing in current docs here:
> Each call to observable.subscribe triggers its own independent for that given Observer.
It seems like a word belongs after "independent", but I'm not sure what, so I just suggested "setup".

And for this one:
> This shows how a subscribe calls are not at all shared among multiple Observers of the same Observable.
My removal of "a" is fixing a typo, but my removal of "at all" is just simplification. I see why you might want to keep it for emphasis.