refactor(*): rename isUnsubscribed to closed
The Observable specification currently in stage 1 has added a `closed` property to observers which is functionally identical to the `isUnsubscribed` property in RxJS (source: https://github.com/zenparsing/es-observable/issues/90). In this commit we rename everywhere `isUnsubscribed` appears with `closed` to continue to be specification compliant.

Closes #1798

BREAKING CHANGE: rename `isUnsubscribed` to `closed`