feat(op): Adds the new .op() method to Observable
Allows you to apply an operator to a given stream, without it needing to
be
on the Observable prototype. This is useful whenever you don't want to
unintentionally leak implementation details or otherwise rely on the
fact
that you're using a particular operator.

It takes the provided operator function and any arbitrary arguments
then internally just calls `operator.apply(this, args)`.

This is solves a similar problem as the TC39 proposed [bind operator
syntax](https://github.com/tc39/proposal-bind-operator)
e.g. `stream::map(i => i + 1)` but that's stage-0 and not supported by
TypeScript.

Example 1: in your unit tests, if you were to patch additional operators
onto the Observable prototype then the application you're testing could
unknowingly be relying on those operators without it importing them
itself.
Your tests would pass because the operator is available in your tests,
but
when you run the app standalone it errors because the operator wasn't
imported!

Example 2: you're writing a library that itself uses RxJS. If you patch
operators onto the prototype, then consumers of your library could
accidentally depend on the fact that the operators you use are
available.