fix(buffer): emit last buffer if source completes
Fix buffer operator not emitting the last buffer when the source
completes. This is closer to rxjs v4* behaviour and matches the
behaviour of the other buffer operators. The _complete method is very
similar to thoses in bufferCount, bufferTime and bufferWhen.

*rxjs v4 will always emit the buffer if the source completes even if the
buffer is empty. This fix only emits if the buffer is non empty.

BREAKING CHANGE:

The `buffer()` operator now emits what's partially buffered when the
source completes. This is closer to rxjs v4* behaviour and matches the
v5 behaviour of the other buffer operators.