docs(DeferObservable.ts, EmptyObservable.ts, FromEventObservable.ts, FromObservable.ts, combineLates
Add output results for the examples in the documentation pertaining to the operators. Change

seqenceEqual.ts 'var code = Observable.from' to 'var code = Rx.Observable.from'. If the current

example is pasted into an editor and executed it will throw an error of 'Observable not defined'.

The addition of Rx. fixes this.

Issue: Examples do not have resulting output #2171