docs(tutorial): Changed values to keys in the examples
It seems that there is a confusion between event.key and event.target.value in the doc.
As "hello world" is typed in an input box, a stream of keypress events will give output:
    "h"
    "e"
    "l"
    "l"
    ...
for event.key and will output:
    "h"
    "he"
    "hel"
    "hell"
    "hello"
    ...
for event.target.value.
The examples seem to imply to want the first the behavior but they are using the second method.
Changed the examples code to use the first method. 

Another possibility would be to correct the example having in mind the second case.