feat(Zone): add support for Zone.js
Add support for Zone.js in a way which should have minimal impact on performance.

Ensures that all callbacks are invoked in the same `Zone` as the `Zone` which was current at
the time of the callback registration with the `Rx`.

-  Add `getZone()` method which returs always `null` if no `Zone` is present or returrns
   the current zone.
 - All places where the `Zone` needs to be intered are quarded to only go through zone if the
   Zone is different from the current `Zone` for performance reasons.

You can read up on Zones in [The Zone Primer](https://docs.google.com/document/d/1F5Ug0jcrm031vhSMJEOgp1l-Is-Vf0UCNDY-LsQtAIY).
Before this change when Rx runs in the Zone environment it will
propagate Zones according to VM Tasks. In the case of Rx this is not
correct behavior because Rx does its own scheduling and the result is
that all callbacks end up running in the same zone as Task zone. This
behavior is mostly correct most of the time, but there are cases which
result in wrong behavior which this change corrects. This change is
needed to enable proper implementation of
https://github.com/angular/angular/issues/13248.