Leading/Trailing throttle and throttleTime configuration (#2465)
* feat(throttle): add leading/trailing configurability

Adds the ability to configure the `throttle` operator via a
configuration object in the same manner as Lodash's throttle.

Currently this defaults to the existing behavior, which is `{ leading:
true, trailing: false }`. In an upcoming major version, I think we
should change this to align with other libraries that have similar
functionality. In particular Lodash, given it's popularity.

NOTE: While working on this I discovered that the leading value is
actually being emitted *after* the durationSelector is called and
subscribed to. This is likely a bug, but I'll file a separate issue.

related #1625

* feat(throttleTime): add leading/trailing configurability

resolves #1625

* chore(throttle/throttleTime-spec): fix diagram text