fix(groupBy): group duration notifiers will now properly unsubscribe and clean up (#2662)
* test(groupBy): Test that GroupBy-durationSelectors are disposed of

Duration selectors where not disposed when the GroupDurationSubscriber's completed

* fix(groupBy): unsubscribe GroupDurationSubscriber after completing the group

The Groups are disposed by the GroupDurationSelector, however the
GroupDurationSubscriber can be subscribed to a different observable
than the group itself. To prevent any unwanted subscriptions to
accumulate over time we need to explicitly unsubscribe after the
first event in GroupDurationSubscriber closes the group.

Fixes #2660

* fix(groupBy): ensures durationSelector subscriptions are cleaned

The subscriptions to the durationSelector would pile up in the
internal subscription list of the GroupBySubscriber. By removing
the GroupDurationSubscriber explicitly from the GroupBySubscriber
we prevent potential OOM exceptions.

Fixes #2661

* refactor(groupBy): Refactor GroupDurationSubscriber to pass-through to underlying group Subject

* refactor(groupBy): simplify durationSelector test