fix: give type args to higherOrder call in src/operator/publishReplay.ts
operator/publishReplay delegates to operators/publishReplay. The call is
correct &mdash; *one* of the three overloads will be called, and it
doesn't matter which &mdash; but
the compiler has to choose one overload. Previously, the call didn't
include type arguments, and the type of the one argument that would be
used for inference, selectorOrScheduler, was of type `any`. So the
compiler chose the first overload with four parameters. However, the
return type of this overload is incorrect, whereas the return type of the
third overload is correct. Provide the type
arguments explicitly so that the last, most general overload gets called. This
results in the the correct return type, `OperatorFunction<T, R>`.

Closes #2991