operator/publishReplay.ts:give type args to higherOrder call
operator/publishReplay delegates to operators/publishReplay. The call is
correct &mdash; *one* of the three overloads will be called, and it
doesn't matter which &mdash; but
the compiler has to choose one overload. This change provides the type
arguments explicitly so that the most general overload gets called. This
results in the the correct return type, `OperatorFunction<T, R>`.