refactor(elementAt): implement in terms of first
Reduces the size of the operator substantially.

NOTE: We might want to consider changing the error we expect from this and first to be something uniform to remove a step there is a small mismatch between EmptyError and ArgumentOutOfRangeError. I understand the semantics, but I think might be able to come up with an Error that suits both