refactory(rxjs): improve hack for XMLHttpRequest usage with TypeScript 2.7
The XMLHttpRequest type is defined to always have the `request`
property. This will lead 2.7 to infer `never` for the else
part of any test of the form `'request' in xhr` as this, as far
as the declarations are concerned, is always true and TypeScript
then infers the else part will `never` run.

This change prevents this inference wihtout removing the type
checking of the other members of xhr such as `responseText`.