fix: remove the root operators.ts because it overshadows operators/package.json
This affects SystemJS compatibility because we now need to add packages configuration for all
entry-points but there seems to be no way to satify everyone.

The operators/package.json file is needed in order to make prod build resolution work, so we
prefer to keep this working and ask SystemJS users to add extra configuration.

The SystemJS config after this change should be:

System.config({
  map: { 'rxjs': '/some/root/path/to/rxjs/' },
  packages: {
    'rxjs': {main: 'index.js', defaultExtension: 'js' },
    'rxjs/ajax': {main: 'index.js', defaultExtension: 'js' },
    'rxjs/operators': {main: 'index.js', defaultExtension: 'js' },
    'rxjs/testing': {main: 'index.js', defaultExtension: 'js' },
    'rxjs/websocket': {main: 'index.js', defaultExtension: 'js' }
  }
});