fix(subscribe): add internal reportError method
If _trySubscribe catches an error and the sink - or one of its
destinations - is stopped and the sink's error method is called, the
error is swallowed - as it cannot be reported via the observer's
callback. This commit adds an internal reportError method to Subscriber.
The method defers to the appropriate error reporting mechanism if a
stopped subscriber is found.

Closes #3461