feat(isObservable): a new method for checking to see if an object is an RxJS Observable
I've seen a lot of code that checks to see if something is an observable by doing an `typeof` or even `instanceOf` check. It's plausible that in the future these tests might break, as they're testing implementation details of the library. It is recommended that people use this `isObservable` method to see if an object is a compatible RxJS Observable.