fix(subscribable): make subscribe() signature match Observable (#4050)
* fix(subscribable): make subscribe() signature match Observable

Fixes #3891

* test: cast string | number to string before concatenating

T of this Observable is string | number, so TypeScript does not allow
using +. The desired behaviour for this function is to concatenate the
string values, so cast to strings. On master, these were incorrectly
inferred as never, which is why this error was not surfaced.

* test: correct type for selector function

The hot() function will create an Observable that emits strings from
marble diagrams (even if the strings contain numbers).
So the type of x should be string, not number.