docs(resources): Add New Section in Docs for External Resources
This is a starting boilerplate to introduce external resources that will
be useful for users when looking for other places to learn. The
boilerplate is rough draft with the goal of figuring out the best way to
manage an up to date list of resources.

I don't want to sit on this yet at the same time have to be careful of
making sure that it is maintained and kept up to date. It can very
quickly become out of date if the content no longer is relevant to the
current version of RxJS.

Thoughts are welcomed. There have been talks already started about
creating a more dynamic resources, similar to Teams. I would like to get
information in front of people in quicker iteration and build that out
on a separate branch using the data that will continue to populate this
markdown file.