docs(scheduler): correct the description of asapScheduler
`asapScheduler` does not use the micro task queue. As far as I know, only `process.nextTick`, native `Promise`, `Object.observe`, `MutationObserver` used the micro task queue. What I found in the code was that it put the task into a custom queue and set up a macro task to execute the tasks of the queue.