feat(async,spawn): add async functions for observables
A [longer version][] of the justification for this change is available
online. In brief, we're missing the observable equivalent of async
functions. Chaining operators like `map`, `filter`, and `catchError`
with observables that yield a single value is as painful as chaining
callbacks on promises. Async functions let us write asynchronous code in
a synchronous style, which is easier to read, write, and comprehend.
Short of native syntax like async/await, we can offer a decorator for
generator functions.

[longer version]: https://medium.com/@thejohnfreeman/escaping-pipeline-hell-38d962f66d31

Fixes #1497