docs(ajax): describe your change...
I have the feeling that since `ajax` is one of the useful operators we have in Rxjs for making requests etc etc it would be nice if the documentation have more examples. Therefore I opened this PR that will add few examples on Stackblitz. Please look here.

I would be happy to hear your feedback since now I am gonna work on that in order to have this documented well for future referenced