fix(WebSocketSubject): close websocket connection attempt on unsubscribe
Fix closing the underlying WebSocket when a WebSocketSubject is
unsubscribed while the WebSocket is in the `CONNECTING` state.
Initial support for closing a WebSocket before it was opened was added in #4446.
However, that pull-request still waits for the connection to eventually
become open before attempting to close it.
This pull-request takes it a step further by closing the websocket as soon as the
subject is unsubscribed.