feat(ReplaySubject): Remove scheduler dependency
The aim here is to simplify `ReplaySubject` by removing a unnecessary feature that pulled in
a lot of scheduler code via `QueueScheduler` and `ObserveOnSubscriber`, which in turn pulled in
`AsyncScheduler`, and a lot of other related code.

Instead, we're going to pull in a new, smaller `TimestampProvider` type, that simply has the necessary
`now` method on it, meaning that `Date` or `performance` could be used as arguments to the `ReplaySubject`
constructor.

BREAKING CHANGE: `ReplaySubject` no longer schedules emissions when a scheduler is provided. If you need that behavior,
please compose in `observeOn` using `pipe`, for example: `new ReplaySubject(2, 3000).pipe(observeOn(asap))`