feat(retry): add config to reset error count on successful emission (#5280)
* feat(retry): add config to reset error count on successful emission

This PR adds the ability to reset the error counter on successful emissions using the `retry`
operator. The current behavior for `retry(n)` is to call error if n errors occurred, regardless of
whether or not they were consecutive. Now one would be able to use `retry(n, true)` to have the
count reset so that only n consecutive errors will cause the observable to fail.

* feat(retry): add config parameter

added overloaded signature to the `retry` operator that accepts a config object

* chore: restore package-lock.json

* chore: appease TypeScript

* chore: revert change to Observable spec