Update resources.md
My playlist is still very much a work in progress, and I'm taking a somewhat controversial approach with introducing subjects before observables, but I think for various reasons some people may learn better this way, especially if they think like me. I do discuss all the tradeoffs there of. I've yet to make the videos about higher order observables, error handling, control flow, and walk-throughs of operators, so if there is any objection to adding my work in progress playlist, I'm happy to close this PR / let it wait until the playlist is more complete - not sure if ya'll review the technical content for correctness before linking to it from here, so I understand if you want to close this PR for now, and hopefully we can revisit it later, but I think the content I've recorded so far could be super helpful to some.