refactor(timeout): Getting rid of _unsubscribeAndRecycle (#5626)
In an effort to simplify the code base, along with the other work to retry, retryWhen, repeat, and repeatWhen, this should help us get rid of `_unsubscribeAndRecycle` and move us toward simplifying Subscriber in the long run. There are obviously some memory trade offs using functions on POJOs over on prototypes, but given other recent memory improvements, I think that is probably not a big deal.